# Optimised Code and Fixed Tests