#### Check List:
- [x] Tests pass (`./test.sh <username>` shows no differences between expected and actual outputs)
- [x] All formatting changes by the build are committed
- [x] Your launch script is named `calculate_average_<username>.sh` (make sure to match casing of your GH user name) and is executable
- [x] Output matches that of `calculate_average_baseline.sh`
* Execution time: 17.636s
* Execution time of reference implementation: 47.717s

Using java version 21.0.1-graal in this shell.
Benchmark 1: ./calculate_average_kuduwa-keshavram.sh 2>&1
  Time (mean ± σ):     17.636 s ±  0.296 s    [User: 51.162 s, System: 14.052 s]
  Range (min … max):   17.256 s … 18.076 s    5 runs
 
Summary
  kuduwa-keshavram: trimmed mean 17.621236375546662, raw times 17.64174988888,17.256375597879998,17.531187930879998,18.076145597879997,17.67601455588

<!--
Thanks for your submission. Please go through the checklist above before submitting your pull request.
Use [x] to mark that the item has been completed.

Please submit only submissions that are taking less than 1 minute.

Please make sure that you have followed the defined rules (https://github.com/gunnarmorling/1brc?tab=readme-ov-file#rules-and-limits).
-->